refactor(libs-modules-auth-api): move auth code to own lib
- Use @Global module for auth & dgraph
