+++ apps/api/src/app/app.module.ts
@@ -1,6 +1,7 @@
-import { AuthModule, InfrastructureModule } from '@codelab/backend'
+import { InfrastructureModule } from '@codelab/backend'
 import { AppModule as AppApiModule } from '@codelab/modules/app-api'
 import { AtomModule } from '@codelab/modules/atom-api'
+import { AuthModule } from '@codelab/modules/auth-api'
 import { LambdaApiModule } from '@codelab/modules/lambda-api'
 import { PageModule } from '@codelab/modules/page-api'
 import { PageElementModule } from '@codelab/modules/page-element-api'
