--- libs/backend/src/application/auth/index.ts
@@ -1,5 +1,0 @@
-export * from './auth.module'
-export * from './auth0'
-export * from './config/auth.tokens'
-export * from './graphql'
-export * from './interfaces/jwt.interface'
