+++ libs/backend/src/application/index.ts
@@ -1,4 +1,3 @@
-export * from './auth'
 export * from './dgraph'
 export * from './graphql'
 export * from './use-case'
