+++ libs/backend/src/infrastructure/dgraph/dgraph.module.ts
@@ -1,8 +1,9 @@
-import { Module } from '@nestjs/common'
+import { Global, Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { dgraphConfig } from './config/dgraph.config'
 import { DGraphService } from './dgraph.service'
 
+@Global()
 @Module({
   imports: [ConfigModule.forFeature(dgraphConfig)],
   providers: [DGraphService],
