+++ libs/modules/app-api/src/app.resolver.ts
@@ -1,5 +1,8 @@
-import type { JwtPayload } from '@codelab/backend'
-import { CurrentUser, GqlAuthGuard } from '@codelab/backend'
+import {
+  CurrentUser,
+  GqlAuthGuard,
+  JwtPayload,
+} from '@codelab/modules/auth-api'
 import { Injectable, UseGuards } from '@nestjs/common'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 import { App } from './app.model'
