+++ libs/modules/app-api/src/use-cases/delete-app/delete-app.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { DeleteAppInput } from './delete-app.input'
 
 export class DeleteAppRequest {
