+++ libs/modules/app-api/src/use-cases/get-app/get-app.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { GetAppInput } from './get-app.input'
 
 export class GetAppRequest {
