+++ libs/modules/app-api/src/use-cases/update-app/update-app.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { UpdateAppInput } from './update-app.input'
 
 export class UpdateAppRequest {
