+++ libs/modules/auth-api/src/config/auth.config.ts
