+++ libs/modules/auth-api/src/index.ts
@@ -1,1 +1,6 @@
-export * from './lib/modules-auth-api.module'
+export * from './auth.module'
+export * from './auth0'
+export * from './config/auth.tokens'
+export * from './graphql'
+export * from './interfaces/jwt.interface'
+export * from './strategy/jwt.strategy'
