+++ libs/modules/auth-api/src/interfaces/jwt.interface.ts
