--- libs/modules/auth-api/src/lib/modules-auth-api.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common'
-
-@Module({
-  controllers: [],
-  providers: [],
-  exports: [],
-})
-export class ModulesAuthApiModule {}
