+++ libs/modules/page-api/src/page.resolver.ts
@@ -1,5 +1,8 @@
-import type { JwtPayload } from '@codelab/backend'
-import { CurrentUser, GqlAuthGuard } from '@codelab/backend'
+import {
+  CurrentUser,
+  GqlAuthGuard,
+  JwtPayload,
+} from '@codelab/modules/auth-api'
 import { PageElementRoot } from '@codelab/modules/page-element-api'
 import { Injectable, UseGuards } from '@nestjs/common'
 import {
