+++ libs/modules/page-api/src/use-cases/get-page-root/get-page-root.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { GetPageRootInput } from './get-page-root.input'
 
 export class GetPageRootRequest {
