+++ libs/modules/page-api/src/use-cases/get-page/get-page.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { GetPageInput } from './get-page.input'
 
 export class GetPageRequest {
