+++ libs/modules/page-api/src/use-cases/get-pages/get-pages.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { GetPagesInput } from './get-pages.input'
 
 export class GetPagesRequest {
