+++ libs/modules/page-api/src/use-cases/update-page/update-page.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { UpdatePageInput } from './update-page.input'
 
 export class UpdatePageRequest {
