+++ libs/modules/page-element-api/src/auth/page-element-guard/page-element-guard.service.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { Injectable } from '@nestjs/common'
 //shortened import causes circular reference and some weird shit happen
 import { GetPageElementOwnerService } from '../../use-cases/get-page-element-owner/get-page-element-owner.service'
