+++ libs/modules/page-element-api/src/page-element.resolver.ts
@@ -1,5 +1,9 @@
-import type { JwtPayload } from '@codelab/backend'
-import { CurrentUser, DeleteResponse, GqlAuthGuard } from '@codelab/backend'
+import { DeleteResponse } from '@codelab/backend'
+import {
+  CurrentUser,
+  GqlAuthGuard,
+  JwtPayload,
+} from '@codelab/modules/auth-api'
 import { Injectable, UseGuards } from '@nestjs/common'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 import { PageElement, PageElementRoot } from './models'
