+++ libs/modules/page-element-api/src/use-cases/create-page-element/create-page-element.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { CreatePageElementInput } from './create-page-element.input'
 
 export class CreatePageElementRequest {
