+++ libs/modules/page-element-api/src/use-cases/delete-page-element/delete-page-element.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { DeletePageElementInput } from './delete-page-element.input'
 
 export class DeletePageElementRequest {
