+++ libs/modules/page-element-api/src/use-cases/get-page-element/get-page-element.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { GetPageElementInput } from './get-page-element.input'
 
 export class GetPageElementRequest {
