+++ libs/modules/page-element-api/src/use-cases/move-page-element/move-page-element.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { MovePageElementInput } from './move-page-element.input'
 
 export class MovePageElementRequest {
