+++ libs/modules/page-element-api/src/use-cases/update-page-element/update-page-element.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { UpdatePageElementInput } from './update-page-element.input'
 
 export class UpdatePageElementRequest {
