+++ libs/modules/user-api/src/use-cases/get-user/get-user.service.ts
@@ -1,4 +1,5 @@
-import { Auth0Service, UseCase } from '@codelab/backend'
+import { UseCase } from '@codelab/backend'
+import { Auth0Service } from '@codelab/modules/auth-api'
 import { Injectable } from '@nestjs/common'
 import { User } from '../../user.model'
 import { GetUserRequest } from './get-user.request'
