+++ libs/modules/user-api/src/use-cases/update-user/update-user.request.ts
@@ -1,4 +1,4 @@
-import { JwtPayload } from '@codelab/backend'
+import { JwtPayload } from '@codelab/modules/auth-api'
 import { UpdateUserInput } from './update-user.input'
 
 export class UpdateUserRequest {
