+++ libs/modules/user-api/src/user.resolver.ts
@@ -1,5 +1,10 @@
-import type { JwtPayload } from '@codelab/backend'
-import { CurrentUser, GqlAuthGuard, GqlRoleGuard, Role } from '@codelab/backend'
+import {
+  CurrentUser,
+  GqlAuthGuard,
+  GqlRoleGuard,
+  JwtPayload,
+  Role,
+} from '@codelab/modules/auth-api'
 import { Injectable, UseGuards } from '@nestjs/common'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 import {
