+++ libs/modules/value-type-api/src/valueType.module.ts
@@ -1,10 +1,8 @@
-import { ApolloClientModule, AuthModule, DGraphModule } from '@codelab/backend'
 import { Module } from '@nestjs/common'
 import { GetValueTypesService } from './use-cases'
 import { ValueTypeResolver } from './valueType.resolver'
 
 @Module({
-  imports: [DGraphModule, AuthModule, ApolloClientModule],
   providers: [GetValueTypesService, ValueTypeResolver],
   exports: [],
 })
